refactor: 게스트 초대 취소 로직 CommandService에서 호출하도록 변경
