+++ src/main/java/umc/cockple/demo/domain/exercise/service/ExerciseGuestService.java
@@ -45,13 +45,7 @@ public class ExerciseGuestService {
         return exerciseConverter.toGuestInviteResponse(savedGuest, exercise);
     }
 
-    public ExerciseCancelDTO.Response cancelGuestInvitation(Long exerciseId, Long guestId, Long memberId) {
-
-        log.info("게스트 초대 취소 시작 - exerciseId: {}, guestId: {}, memberId: {}", exerciseId, guestId, memberId);
-
-        Exercise exercise = findExerciseOrThrow(exerciseId);
-        Member member = findMemberOrThrow(memberId);
-        Guest guest = findGuestOrThrow(guestId);
+    public ExerciseCancelDTO.Response cancelGuestInvitation(Exercise exercise, Guest guest, Member member) {
         exerciseValidator.validateCancelGuestInvitation(exercise, guest, member);
 
         exercise.removeGuest(guest);
@@ -60,25 +54,8 @@ public class ExerciseGuestService {
 
         exerciseRepository.save(exercise);
 
-        log.info("게스트 초대 취소 완료 - exerciseId: {}, guestId: {}, memberId: {}", exerciseId, guestId, memberId);
+        log.info("게스트 초대 취소 완료 - exerciseId: {}, guestId: {}, memberId: {}", exercise.getId(), guest.getId(), member.getId());
 
         return exerciseConverter.toCancelResponse(exercise, guest);
-    }
-
-    // ========== 조회 메서드 ==========
-
-    private Exercise findExerciseOrThrow(Long exerciseId) {
-        return exerciseRepository.findById(exerciseId)
-                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.EXERCISE_NOT_FOUND));
-    }
-
-    private Member findMemberOrThrow(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.MEMBER_NOT_FOUND));
-    }
-
-    private Guest findGuestOrThrow(Long guestId) {
-        return guestRepository.findById(guestId)
-                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.GUEST_NOT_FOUND));
     }
 }
